requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Vector Database (Pinecone cloud - free tier: 100K vectors)
pinecone>=5.0.0

# Document Processing
pdfplumber>=0.10.0
ebooklib>=0.18
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# LLM
groq>=0.4.0

# LangChain (LCEL routing + recursive chunking)
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-text-splitters>=0.2.0

# Configuration
python-dotenv>=1.0.0

# Utilities
tiktoken>=0.5.0

# Embeddings and Re-ranking (Cohere free tier)
cohere>=5.0.0

# Evaluation (RAGAS + Gemini as judge)
ragas>=0.2.0
google-genai>=1.0.0
datasets>=2.0.0

# Hybrid Retrieval
rank-bm25>=0.2.2

# Authentication
python-jose[cryptography]>=3.3.0

# Database (PostgreSQL)
asyncpg>=0.29.0

# Research Agent
tavily-python

# MCP Server
mcp[cli]>=1.2.0

# A2A Protocol
a2a-sdk[http-server]>=0.3.0
